+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 Each tagged release MUST have a section 2 heading starting at the time of release `## TAG-NAME...` or Gitlab release notes will be missed.
 
+## v0.30.1
+- Development environment (`mw docker`)
+  - Removed faulty `develop` from default `XDEBUG_MODE` (T366986, T420913)
+
 ## v0.30.0
 
 - Development environment (`mw docker`)